Make it work in Ruby 2.4
My assumption was, that Ruby 2.4 would trigger `Warning::warn` for all calls to `Kernel::warn` and for all internal warnings. Apparently this was not correct. Only the internal warnings are routed through `Warning::warn`. Consequently we need to load `StructuredWarnings::Kernel` for all Ruby versions, so that we can extend user land warnings as well.